Done CPP_00

+++ CPP_00/ex00/megaphone.cpp
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   megaphone.cpp                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jobject <jobject@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/12/13 12:19:12 by jobject           #+#    #+#             */
+/*   Updated: 2021/12/13 12:19:12 by jobject          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include <iostream>
 #include <string>
 
@@ -11,7 +23,7 @@ int main(int argc, char **argv)
     }
     for (int i = 1; i < argc; i++) {
         std :: string av(argv[i]);
-        for (int j = 0; j < av.size(); j++)
+        for (size_t j = 0; j < av.size(); j++)
             std :: cout << (char) std :: toupper(av[j]);
     }
     std :: cout << std :: endl;
